Change on the metodology of gathering the packets
In order to be able to capture all the packets that are being send/received, last version used forks, to avoid delay between saving files. But that perspective doesn't alow us to count the up/down every sec info and send it to the main process, which could be done with a queue or by writing a file with the result(which will lead us to more dead time).
That's why I took the decision of stop writing multiple files, and do all the gathering work before analyzing and saving the data.